pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "iil-reflex"
version = "0.5.0"
description = "REFLEX — Evidence-based UI development with LLM domain agent, UC dialog, permission testing, and dev cycle orchestration"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    { name = "Achim Dehnert", email = "achim@iil.gmbh" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Quality Assurance",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "iil-promptfw>=0.7.0,<1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
llm = ["litellm>=1.40"]
aifw = ["iil-aifw>=0.9.0"]
playwright = ["playwright>=1.40"]
metrics = ["psycopg[binary]>=3.1"]
web = [
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "pdfplumber>=0.11",
]
dev = [
    "pytest>=8.0",
    "pytest-cov",
    "ruff>=0.4",
    "litellm>=1.40",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
]
all = [
    "iil-reflex[llm,aifw,playwright,web,dev]",
]

[project.scripts]
reflex = "reflex.__main__:main"

[project.urls]
Homepage = "https://github.com/achimdehnert/iil-reflex"
Repository = "https://github.com/achimdehnert/iil-reflex"
Documentation = "https://knowledge.iil.pet/doc/adr-162-reflex"

[tool.hatch.build.targets.sdist]
exclude = [".windsurf", ".github"]

[tool.hatch.build.targets.wheel]
packages = ["reflex"]


[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"reflex/dashboard.py" = ["E501"]  # inline HTML strings
"reflex/dashboard/template.py" = ["E501"]  # inline HTML strings

[tool.pytest.ini_options]
testpaths = ["tests"]
